updated game class
